+++ language/en/postlove.php
@@ -62,7 +62,8 @@ $lang = array_merge($lang, array(
 		1	=> 'Once in total',
 		2	=> '%d times in total',
 	),
-	'POSTLOVE_HIDE' 			=> 'Hide Like icons and summaries',
+	'POSTLOVE_HIDE' 			=> 'Hide the Like button on posts',
+	'POSTLOVE_HIDE_SUM'	=> 'Hide most-liked-posts summaries',
 	'ACL_U_POSTLOVE'			=> 'Post Love: Can like posts',
 	'ACL_U_POSTLOVE_SUMMARY'	=> 'Post Love: Can see most liked posts summary',